requirements.txt
aws-cdk.core>=1.133.0
aws-cdk.assertions>=1.133.0
aws-cdk.aws-ec2>=1.133.0
aws-cdk.aws-iam>=1.133.0
aws-cdk.aws-opensearchservice>=1.133.0
aws-cdk.aws-s3>=1.133.0
aws-cdk.aws-sagemaker>=1.133.0
